test(scim): split scim.test.ts to fix CI flaky timeout
Split the monolithic scim.test.ts (2357 lines, 54 tests) into 3 files
so vitest runs them in separate workers, reducing resource contention
on CI. Also bump testTimeout to 10s in the scim vitest config.

- scim.test.ts: ServiceProviderConfig, Schemas, ResourceTypes, POST/PUT Users
- scim-patch.test.ts: PATCH /scim/v2/users tests
- scim-users.test.ts: GET/DELETE Users + Default SCIM provider tests
