+++ packages/scim/src/scim.test.ts
@@ -1085,1273 +1085,4 @@ describe("SCIM", () => {
 			);
 		});
 	});
-
-	describe("PATCH /scim/v2/users", () => {
-		it.for([
-			"replace",
-			"add",
-		])("should partially update a user resource with %s", async (op) => {
-			const { auth, getSCIMToken } = createTestInstance();
-			const scimToken = await getSCIMToken();
-
-			const user = await auth.api.createSCIMUser({
-				body: {
-					userName: "the-username",
-					name: {
-						formatted: "Juan Perez",
-					},
-					emails: [{ value: "primary-email@test.com", primary: true }],
-				},
-				headers: {
-					authorization: `Bearer ${scimToken}`,
-				},
-			});
-
-			expect(user).toBeTruthy();
-			expect(user.externalId).toBe("the-username");
-			expect(user.userName).toBe("primary-email@test.com");
-			expect(user.name.formatted).toBe("Juan Perez");
-			expect(user.emails[0]?.value).toBe("primary-email@test.com");
-
-			await auth.api.patchSCIMUser({
-				params: {
-					userId: user.id,
-				},
-				body: {
-					schemas: ["urn:ietf:params:scim:api:messages:2.0:PatchOp"],
-					Operations: [
-						{ op: op, path: "/externalId", value: "external-username" },
-						{ op: op, path: "/userName", value: "other-username" },
-						{ op: op, path: "/name/givenName", value: "Daniel" },
-					],
-				},
-				headers: {
-					authorization: `Bearer ${scimToken}`,
-				},
-			});
-
-			const updatedUser = await auth.api.getSCIMUser({
-				params: {
-					userId: user.id,
-				},
-				headers: {
-					authorization: `Bearer ${scimToken}`,
-				},
-			});
-
-			expect(updatedUser).toMatchObject({
-				active: true,
-				displayName: "Daniel Perez",
-				emails: [
-					{
-						primary: true,
-						value: "other-username",
-					},
-				],
-				externalId: "external-username",
-				id: expect.any(String),
-				meta: expect.objectContaining({
-					created: expect.any(Date),
-					lastModified: expect.any(Date),
-					location: expect.stringContaining("/api/auth/scim/v2/Users/"),
-					resourceType: "User",
-				}),
-				name: {
-					formatted: "Daniel Perez",
-				},
-				schemas: expect.arrayContaining([
-					"urn:ietf:params:scim:schemas:core:2.0:User",
-				]),
-				userName: "other-username",
-			});
-		});
-
-		it("should partially update a user resource with mixed operations", async () => {
-			const { auth, getSCIMToken } = createTestInstance();
-			const scimToken = await getSCIMToken();
-
-			const user = await auth.api.createSCIMUser({
-				body: {
-					userName: "the-username",
-					name: {
-						formatted: "Juan Perez",
-					},
-					emails: [{ value: "primary-email@test.com", primary: true }],
-				},
-				headers: {
-					authorization: `Bearer ${scimToken}`,
-				},
-			});
-
-			expect(user).toBeTruthy();
-			expect(user.externalId).toBe("the-username");
-			expect(user.userName).toBe("primary-email@test.com");
-			expect(user.name.formatted).toBe("Juan Perez");
-			expect(user.emails[0]?.value).toBe("primary-email@test.com");
-
-			await auth.api.patchSCIMUser({
-				params: {
-					userId: user.id,
-				},
-				body: {
-					schemas: ["urn:ietf:params:scim:api:messages:2.0:PatchOp"],
-					Operations: [
-						{ op: "add", path: "/externalId", value: "external-username" },
-						{ op: "replace", path: "/userName", value: "other-username" },
-						{ op: "add", path: "/name/formatted", value: "Daniel Lopez" },
-					],
-				},
-				headers: {
-					authorization: `Bearer ${scimToken}`,
-				},
-			});
-
-			const updatedUser = await auth.api.getSCIMUser({
-				params: {
-					userId: user.id,
-				},
-				headers: {
-					authorization: `Bearer ${scimToken}`,
-				},
-			});
-
-			expect(updatedUser).toMatchObject({
-				active: true,
-				displayName: "Daniel Lopez",
-				emails: [
-					{
-						primary: true,
-						value: "other-username",
-					},
-				],
-				externalId: "external-username",
-				id: expect.any(String),
-				meta: expect.objectContaining({
-					created: expect.any(Date),
-					lastModified: expect.any(Date),
-					location: expect.stringContaining("/api/auth/scim/v2/Users/"),
-					resourceType: "User",
-				}),
-				name: {
-					formatted: "Daniel Lopez",
-				},
-				schemas: expect.arrayContaining([
-					"urn:ietf:params:scim:schemas:core:2.0:User",
-				]),
-				userName: "other-username",
-			});
-		});
-
-		it.for([
-			"replace",
-			"add",
-		])("should partially update multiple name sub-attributes with %s", async (op) => {
-			const { auth, getSCIMToken } = createTestInstance();
-			const scimToken = await getSCIMToken();
-
-			const user = await auth.api.createSCIMUser({
-				body: {
-					userName: "sub-attribute-test-user",
-					name: {
-						formatted: "Original Name",
-					},
-				},
-				headers: {
-					authorization: `Bearer ${scimToken}`,
-				},
-			});
-
-			await auth.api.patchSCIMUser({
-				params: { userId: user.id },
-				body: {
-					schemas: ["urn:ietf:params:scim:api:messages:2.0:PatchOp"],
-					Operations: [
-						{ op: op, path: "/name/givenName", value: "Updated" },
-						{ op: op, path: "/name/familyName", value: "Value" },
-					],
-				},
-				headers: {
-					authorization: `Bearer ${scimToken}`,
-				},
-			});
-
-			const updatedUser = await auth.api.getSCIMUser({
-				params: { userId: user.id },
-				headers: {
-					authorization: `Bearer ${scimToken}`,
-				},
-			});
-
-			expect(updatedUser.name.formatted).toBe("Updated Value");
-		});
-
-		it.for([
-			"replace",
-			"add",
-		])("should %s nested object values with path prefix", async (op) => {
-			const { auth, getSCIMToken } = createTestInstance();
-			const scimToken = await getSCIMToken();
-
-			const user = await auth.api.createSCIMUser({
-				body: {
-					userName: "nested-test-user",
-					name: { formatted: "Original Name" },
-				},
-				headers: {
-					authorization: `Bearer ${scimToken}`,
-				},
-			});
-
-			await auth.api.patchSCIMUser({
-				params: { userId: user.id },
-				body: {
-					schemas: ["urn:ietf:params:scim:api:messages:2.0:PatchOp"],
-					Operations: [
-						{
-							op: op,
-							path: "name",
-							value: { givenName: "Nested" },
-						},
-						{
-							op: op,
-							path: "name",
-							value: { familyName: "User" },
-						},
-						{
-							op: op,
-							path: "userName",
-							value: "nested-test-user-updated",
-						},
-					],
-				},
-				headers: {
-					authorization: `Bearer ${scimToken}`,
-				},
-			});
-
-			const updatedUser = await auth.api.getSCIMUser({
-				params: { userId: user.id },
-				headers: {
-					authorization: `Bearer ${scimToken}`,
-				},
-			});
-
-			expect(updatedUser.name.formatted).toBe("Nested User");
-			expect(updatedUser.displayName).toBe("Nested User");
-			expect(updatedUser.userName).toBe("nested-test-user-updated");
-		});
-
-		it.for([
-			"replace",
-			"add",
-		])("should support operations without explicit path with %s", async (op) => {
-			const { auth, getSCIMToken } = createTestInstance();
-			const scimToken = await getSCIMToken();
-
-			const user = await auth.api.createSCIMUser({
-				body: {
-					userName: "no-path-test-user",
-				},
-				headers: {
-					authorization: `Bearer ${scimToken}`,
-				},
-			});
-
-			await auth.api.patchSCIMUser({
-				params: { userId: user.id },
-				body: {
-					schemas: ["urn:ietf:params:scim:api:messages:2.0:PatchOp"],
-					Operations: [
-						{
-							op: op,
-							value: {
-								name: { formatted: "No Path Name" },
-								userName: "Username",
-							},
-						},
-					],
-				},
-				headers: {
-					authorization: `Bearer ${scimToken}`,
-				},
-			});
-
-			const updatedUser = await auth.api.getSCIMUser({
-				params: { userId: user.id },
-				headers: {
-					authorization: `Bearer ${scimToken}`,
-				},
-			});
-
-			expect(updatedUser.name.formatted).toBe("No Path Name");
-			expect(updatedUser.userName).toBe("username");
-		});
-
-		it("should support dot notation in paths", async () => {
-			const { auth, getSCIMToken } = createTestInstance();
-			const scimToken = await getSCIMToken();
-
-			const user = await auth.api.createSCIMUser({
-				body: {
-					userName: "dot-notation-user",
-					name: { formatted: "Original Name" },
-				},
-				headers: {
-					authorization: `Bearer ${scimToken}`,
-				},
-			});
-
-			await auth.api.patchSCIMUser({
-				params: { userId: user.id },
-				body: {
-					schemas: ["urn:ietf:params:scim:api:messages:2.0:PatchOp"],
-					Operations: [
-						{ op: "replace", path: "name.familyName", value: "Dot" },
-						{ op: "add", path: "name.givenName", value: "User" },
-						{ op: "add", path: "userName", value: "Username" },
-					],
-				},
-				headers: {
-					authorization: `Bearer ${scimToken}`,
-				},
-			});
-
-			const updatedUser = await auth.api.getSCIMUser({
-				params: { userId: user.id },
-				headers: {
-					authorization: `Bearer ${scimToken}`,
-				},
-			});
-
-			expect(updatedUser.name.formatted).toBe("User Dot");
-			expect(updatedUser.userName).toBe("username");
-		});
-
-		it.for([
-			"replace",
-			"add",
-		])("should handle %s operation case-insensitively", async (op) => {
-			const { auth, getSCIMToken } = createTestInstance();
-			const scimToken = await getSCIMToken();
-
-			const user = await auth.api.createSCIMUser({
-				body: {
-					userName: "user-case-insensitive",
-					name: { formatted: "Original" },
-				},
-				headers: {
-					authorization: `Bearer ${scimToken}`,
-				},
-			});
-
-			await auth.api.patchSCIMUser({
-				params: { userId: user.id },
-				body: {
-					schemas: ["urn:ietf:params:scim:api:messages:2.0:PatchOp"],
-					Operations: [
-						{
-							op: op.toUpperCase(),
-							path: "name.formatted",
-							value: "user-case",
-						},
-					],
-				},
-				headers: {
-					authorization: `Bearer ${scimToken}`,
-				},
-			});
-
-			const updatedUser = await auth.api.getSCIMUser({
-				params: { userId: user.id },
-				headers: {
-					authorization: `Bearer ${scimToken}`,
-				},
-			});
-
-			expect(updatedUser.name.formatted).toBe("user-case");
-		});
-
-		it("should skip add operation when value already exists", async () => {
-			const { auth, getSCIMToken } = createTestInstance();
-			const scimToken = await getSCIMToken();
-
-			const user = await auth.api.createSCIMUser({
-				body: {
-					userName: "add-same-info-user",
-					name: { formatted: "Existing Name" },
-				},
-				headers: {
-					authorization: `Bearer ${scimToken}`,
-				},
-			});
-
-			const patchUser = () =>
-				auth.api.patchSCIMUser({
-					params: { userId: user.id },
-					body: {
-						schemas: ["urn:ietf:params:scim:api:messages:2.0:PatchOp"],
-						Operations: [
-							{ op: "add", path: "/name/formatted", value: "Existing Name" },
-						],
-					},
-					headers: {
-						authorization: `Bearer ${scimToken}`,
-					},
-				});
-
-			await expect(patchUser()).rejects.toThrowError(
-				expect.objectContaining({
-					message: "No valid fields to update",
-					body: {
-						detail: "No valid fields to update",
-						schemas: ["urn:ietf:params:scim:api:messages:2.0:Error"],
-						status: "400",
-					},
-				}),
-			);
-		});
-
-		it.for([
-			"replace",
-			"add",
-		])("should ignore %s on non-existing path", async (op) => {
-			const { auth, getSCIMToken } = createTestInstance();
-			const scimToken = await getSCIMToken();
-
-			const user = await auth.api.createSCIMUser({
-				body: {
-					userName: "non-existing-path",
-					name: { formatted: "Original Name" },
-				},
-				headers: {
-					authorization: `Bearer ${scimToken}`,
-				},
-			});
-
-			const patchUser = () =>
-				auth.api.patchSCIMUser({
-					params: { userId: user.id },
-					body: {
-						schemas: ["urn:ietf:params:scim:api:messages:2.0:PatchOp"],
-						Operations: [
-							{ op: op, path: "/nonExistentField", value: "Some Value" },
-						],
-					},
-					headers: {
-						authorization: `Bearer ${scimToken}`,
-					},
-				});
-
-			await expect(patchUser()).rejects.toThrowError(
-				expect.objectContaining({
-					message: "No valid fields to update",
-					body: {
-						detail: "No valid fields to update",
-						schemas: ["urn:ietf:params:scim:api:messages:2.0:Error"],
-						status: "400",
-					},
-				}),
-			);
-		});
-
-		it("should ignore non-existing operation", async () => {
-			const { auth, getSCIMToken } = createTestInstance();
-			const scimToken = await getSCIMToken();
-
-			const user = await auth.api.createSCIMUser({
-				body: {
-					userName: "non-existing-operation",
-				},
-				headers: {
-					authorization: `Bearer ${scimToken}`,
-				},
-			});
-
-			const patchUser = () =>
-				auth.api.patchSCIMUser({
-					params: { userId: user.id },
-					body: {
-						schemas: ["urn:ietf:params:scim:api:messages:2.0:PatchOp"],
-						Operations: [
-							{ op: "update", path: "userName", value: "Some Value" },
-						],
-					},
-					headers: {
-						authorization: `Bearer ${scimToken}`,
-					},
-				});
-
-			await expect(patchUser()).rejects.toThrowError(
-				expect.objectContaining({
-					body: {
-						code: "VALIDATION_ERROR",
-						message:
-							'[body.Operations.0.op] Invalid option: expected one of "replace"|"add"|"remove"',
-					},
-				}),
-			);
-		});
-
-		it("should return not found for missing users", async () => {
-			const { auth, getSCIMToken } = createTestInstance();
-			const scimToken = await getSCIMToken();
-
-			const patchUser = () =>
-				auth.api.patchSCIMUser({
-					params: {
-						userId: "missing",
-					},
-					body: {
-						schemas: ["urn:ietf:params:scim:api:messages:2.0:PatchOp"],
-						Operations: [
-							{
-								op: "replace",
-								path: "/externalId",
-								value: "external-username",
-							},
-						],
-					},
-					headers: {
-						authorization: `Bearer ${scimToken}`,
-					},
-				});
-
-			await expect(patchUser()).rejects.toThrowError(
-				expect.objectContaining({
-					message: "User not found",
-					body: {
-						detail: "User not found",
-						schemas: ["urn:ietf:params:scim:api:messages:2.0:Error"],
-						status: "404",
-					},
-				}),
-			);
-		});
-
-		it("should fail on invalid updates", async () => {
-			const { auth, getSCIMToken } = createTestInstance();
-			const scimToken = await getSCIMToken();
-
-			const user = await auth.api.createSCIMUser({
-				body: {
-					userName: "the-username",
-				},
-				headers: {
-					authorization: `Bearer ${scimToken}`,
-				},
-			});
-
-			const patchUser = () =>
-				auth.api.patchSCIMUser({
-					params: {
-						userId: user.id,
-					},
-					body: {
-						schemas: ["urn:ietf:params:scim:api:messages:2.0:PatchOp"],
-						Operations: [],
-					},
-					headers: {
-						authorization: `Bearer ${scimToken}`,
-					},
-				});
-
-			await expect(patchUser()).rejects.toThrowError(
-				expect.objectContaining({
-					message: "No valid fields to update",
-					body: {
-						detail: "No valid fields to update",
-						schemas: ["urn:ietf:params:scim:api:messages:2.0:Error"],
-						status: "400",
-					},
-				}),
-			);
-		});
-
-		it("should not allow anonymous access", async () => {
-			const { auth } = createTestInstance();
-
-			const patchUser = async () => {
-				await auth.api.patchSCIMUser({
-					params: {
-						userId: "missing",
-					},
-					body: {
-						schemas: ["urn:ietf:params:scim:api:messages:2.0:PatchOp"],
-						Operations: [
-							{
-								op: "replace",
-								path: "/externalId",
-								value: "external-username",
-							},
-						],
-					},
-				});
-			};
-
-			await expect(patchUser()).rejects.toThrowError(
-				expect.objectContaining({
-					message: "SCIM token is required",
-					body: {
-						detail: "SCIM token is required",
-						schemas: ["urn:ietf:params:scim:api:messages:2.0:Error"],
-						status: "401",
-					},
-				}),
-			);
-		});
-	});
-
-	describe("GET /scim/v2/Users", () => {
-		it("should return the list of users", async () => {
-			const { auth, getSCIMToken } = createTestInstance();
-			const scimToken = await getSCIMToken();
-
-			const createUser = (userName: string) => {
-				return auth.api.createSCIMUser({
-					body: {
-						userName,
-					},
-					headers: {
-						authorization: `Bearer ${scimToken}`,
-					},
-				});
-			};
-
-			const [userA, userB] = await Promise.all([
-				createUser("user-a"),
-				createUser("user-b"),
-			]);
-
-			const users = await auth.api.listSCIMUsers({
-				headers: {
-					authorization: `Bearer ${scimToken}`,
-				},
-			});
-
-			expect(users).toMatchObject({
-				itemsPerPage: 2,
-				schemas: ["urn:ietf:params:scim:api:messages:2.0:ListResponse"],
-				startIndex: 1,
-				totalResults: 2,
-				Resources: [userA, userB],
-			});
-		});
-
-		it("should return an empty list when no users have been provisioned or belong to the organization", async () => {
-			const { auth, getSCIMToken, registerOrganization } = createTestInstance();
-			const scimToken = await getSCIMToken();
-
-			const createUser = (userName: string, scimToken: string) => {
-				return auth.api.createSCIMUser({
-					body: {
-						userName,
-					},
-					headers: {
-						authorization: `Bearer ${scimToken}`,
-					},
-				});
-			};
-
-			const listUsers = (scimToken: string) => {
-				return auth.api.listSCIMUsers({
-					headers: {
-						authorization: `Bearer ${scimToken}`,
-					},
-				});
-			};
-
-			const users = await listUsers(scimToken);
-
-			expect(users).toMatchObject({
-				itemsPerPage: 0,
-				schemas: ["urn:ietf:params:scim:api:messages:2.0:ListResponse"],
-				startIndex: 1,
-				totalResults: 0,
-				Resources: [],
-			});
-
-			const [organizationA, organizationB] = await Promise.all([
-				registerOrganization("org-a"),
-				registerOrganization("org-b"),
-			]);
-
-			const [scimTokenOrgA, scimTokenOrgB] = await Promise.all([
-				getSCIMToken("provider-org-a", organizationA?.id),
-				getSCIMToken("provider-org-b", organizationB?.id),
-			]);
-
-			await createUser("user-a", scimTokenOrgA);
-			const orgBUsers = await listUsers(scimTokenOrgB);
-
-			expect(orgBUsers).toMatchObject({
-				itemsPerPage: 0,
-				schemas: ["urn:ietf:params:scim:api:messages:2.0:ListResponse"],
-				startIndex: 1,
-				totalResults: 0,
-				Resources: [],
-			});
-		});
-
-		it("should only allow access to users that belong to the same provider", async () => {
-			const { auth, getSCIMToken } = createTestInstance();
-			const [scimTokenProviderA, scimTokenProviderB] = await Promise.all([
-				getSCIMToken("provider-a"),
-				getSCIMToken("provider-b"),
-			]);
-
-			const createUser = (userName: string, scimToken: string) => {
-				return auth.api.createSCIMUser({
-					body: {
-						userName,
-					},
-					headers: {
-						authorization: `Bearer ${scimToken}`,
-					},
-				});
-			};
-
-			const listUsers = (scimToken: string) => {
-				return auth.api.listSCIMUsers({
-					headers: {
-						authorization: `Bearer ${scimToken}`,
-					},
-				});
-			};
-
-			const [userA, userB, userC] = await Promise.all([
-				createUser("user-a", scimTokenProviderB),
-				createUser("user-b", scimTokenProviderA),
-				createUser("user-c", scimTokenProviderB),
-			]);
-
-			const [usersProviderA, usersProviderB] = await Promise.all([
-				listUsers(scimTokenProviderA),
-				listUsers(scimTokenProviderB),
-			]);
-
-			expect(usersProviderA).toMatchObject({
-				itemsPerPage: 1,
-				schemas: ["urn:ietf:params:scim:api:messages:2.0:ListResponse"],
-				startIndex: 1,
-				totalResults: 1,
-				Resources: [userB],
-			});
-
-			expect(usersProviderB).toMatchObject({
-				itemsPerPage: 2,
-				schemas: ["urn:ietf:params:scim:api:messages:2.0:ListResponse"],
-				startIndex: 1,
-				totalResults: 2,
-				Resources: [userA, userC],
-			});
-		});
-
-		it("should only allow access to users that belong to the same provider and organization", async () => {
-			const { auth, getSCIMToken, registerOrganization } = createTestInstance();
-			const [organizationA, organizationB] = await Promise.all([
-				registerOrganization("org:a"),
-				registerOrganization("org:b"),
-			]);
-
-			const [scimTokenProviderA, scimTokenProviderB] = await Promise.all([
-				getSCIMToken("provider-a", organizationA?.id),
-				getSCIMToken("provider-b", organizationB?.id),
-			]);
-
-			const createUser = (userName: string, scimToken: string) => {
-				return auth.api.createSCIMUser({
-					body: {
-						userName,
-					},
-					headers: {
-						authorization: `Bearer ${scimToken}`,
-					},
-				});
-			};
-
-			const listUsers = (scimToken: string) => {
-				return auth.api.listSCIMUsers({
-					headers: {
-						authorization: `Bearer ${scimToken}`,
-					},
-				});
-			};
-
-			const [userA, userB, userC] = await Promise.all([
-				createUser("user-a", scimTokenProviderB),
-				createUser("user-b", scimTokenProviderA),
-				createUser("user-c", scimTokenProviderB),
-			]);
-
-			const [usersProviderA, usersProviderB] = await Promise.all([
-				listUsers(scimTokenProviderA),
-				listUsers(scimTokenProviderB),
-			]);
-
-			expect(usersProviderA).toMatchObject({
-				itemsPerPage: 1,
-				schemas: ["urn:ietf:params:scim:api:messages:2.0:ListResponse"],
-				startIndex: 1,
-				totalResults: 1,
-				Resources: [userB],
-			});
-
-			expect(usersProviderB).toMatchObject({
-				itemsPerPage: 2,
-				schemas: ["urn:ietf:params:scim:api:messages:2.0:ListResponse"],
-				startIndex: 1,
-				totalResults: 2,
-				Resources: [userA, userC],
-			});
-		}, 10000);
-
-		it("should filter the list of users", async () => {
-			const { auth, getSCIMToken } = createTestInstance();
-			const scimToken = await getSCIMToken();
-
-			const createUser = (userName: string) => {
-				return auth.api.createSCIMUser({
-					body: {
-						userName,
-					},
-					headers: {
-						authorization: `Bearer ${scimToken}`,
-					},
-				});
-			};
-
-			const [userA] = await Promise.all([
-				createUser("user-a"),
-				createUser("user-b"),
-				createUser("user-c"),
-			]);
-
-			const users = await auth.api.listSCIMUsers({
-				query: {
-					filter: 'userName eq "user-A"',
-				},
-				headers: {
-					authorization: `Bearer ${scimToken}`,
-				},
-			});
-
-			expect(users).toMatchObject({
-				itemsPerPage: 1,
-				schemas: ["urn:ietf:params:scim:api:messages:2.0:ListResponse"],
-				startIndex: 1,
-				totalResults: 1,
-				Resources: [userA],
-			});
-		});
-
-		it("should not allow anonymous access", async () => {
-			const { auth } = createTestInstance();
-
-			const getUsers = async () => {
-				await auth.api.listSCIMUsers();
-			};
-
-			await expect(getUsers()).rejects.toThrowError(
-				expect.objectContaining({
-					message: "SCIM token is required",
-					body: {
-						detail: "SCIM token is required",
-						schemas: ["urn:ietf:params:scim:api:messages:2.0:Error"],
-						status: "401",
-					},
-				}),
-			);
-		});
-	});
-
-	describe("GET /scim/v2/Users/:userId", () => {
-		it("should return a single user resource", async () => {
-			const { auth, getSCIMToken } = createTestInstance();
-			const scimToken = await getSCIMToken();
-
-			const newUser = await auth.api.createSCIMUser({
-				body: {
-					userName: "the-username",
-				},
-				headers: {
-					authorization: `Bearer ${scimToken}`,
-				},
-			});
-
-			const retrievedUser = await auth.api.getSCIMUser({
-				params: {
-					userId: newUser.id,
-				},
-				headers: {
-					authorization: `Bearer ${scimToken}`,
-				},
-			});
-
-			expect(newUser).toEqual(retrievedUser);
-		});
-
-		it("should only allow access to users that belong to the same provider", async () => {
-			const { auth, getSCIMToken } = createTestInstance();
-
-			const [scimTokenProviderA, scimTokenProviderB] = await Promise.all([
-				getSCIMToken("provider-a"),
-				getSCIMToken("provider-b"),
-			]);
-
-			const createUser = (userName: string, scimToken: string) => {
-				return auth.api.createSCIMUser({
-					body: {
-						userName,
-					},
-					headers: {
-						authorization: `Bearer ${scimToken}`,
-					},
-				});
-			};
-
-			const getUser = (userId: string, scimToken: string) => {
-				return auth.api.getSCIMUser({
-					params: {
-						userId,
-					},
-					headers: {
-						authorization: `Bearer ${scimToken}`,
-					},
-				});
-			};
-
-			const [userA, userB] = await Promise.all([
-				createUser("user-a", scimTokenProviderB),
-				createUser("user-b", scimTokenProviderA),
-			]);
-
-			const retrievedUserB = await getUser(userB.id, scimTokenProviderA);
-			expect(retrievedUserB).toEqual(userB);
-
-			await expect(getUser(userB.id, scimTokenProviderB)).rejects.toThrowError(
-				expect.objectContaining({
-					message: "User not found",
-					body: {
-						detail: "User not found",
-						schemas: ["urn:ietf:params:scim:api:messages:2.0:Error"],
-						status: "404",
-					},
-				}),
-			);
-
-			const retrievedUserA = await getUser(userA.id, scimTokenProviderB);
-			expect(retrievedUserA).toEqual(userA);
-
-			await expect(getUser(userA.id, scimTokenProviderA)).rejects.toThrowError(
-				expect.objectContaining({
-					message: "User not found",
-					body: {
-						detail: "User not found",
-						schemas: ["urn:ietf:params:scim:api:messages:2.0:Error"],
-						status: "404",
-					},
-				}),
-			);
-		});
-
-		it("should only allow access to users that belong to the same provider and organization", async () => {
-			const { auth, registerOrganization, getSCIMToken } = createTestInstance();
-			const [organizationA, organizationB] = await Promise.all([
-				registerOrganization("org-a"),
-				registerOrganization("org-b"),
-			]);
-
-			const [scimTokenProviderA, scimTokenProviderB] = await Promise.all([
-				getSCIMToken("provider-a", organizationA?.id),
-				getSCIMToken("provider-b", organizationB?.id),
-			]);
-
-			const createUser = (userName: string, scimToken: string) => {
-				return auth.api.createSCIMUser({
-					body: {
-						userName,
-					},
-					headers: {
-						authorization: `Bearer ${scimToken}`,
-					},
-				});
-			};
-
-			const getUser = (userId: string, scimToken: string) => {
-				return auth.api.getSCIMUser({
-					params: {
-						userId,
-					},
-					headers: {
-						authorization: `Bearer ${scimToken}`,
-					},
-				});
-			};
-
-			const [userA, userB] = await Promise.all([
-				createUser("user-a", scimTokenProviderB),
-				createUser("user-b", scimTokenProviderA),
-			]);
-
-			const retrievedUserB = await getUser(userB.id, scimTokenProviderA);
-			expect(retrievedUserB).toEqual(userB);
-
-			await expect(getUser(userB.id, scimTokenProviderB)).rejects.toThrowError(
-				expect.objectContaining({
-					message: "User not found",
-					body: {
-						detail: "User not found",
-						schemas: ["urn:ietf:params:scim:api:messages:2.0:Error"],
-						status: "404",
-					},
-				}),
-			);
-
-			const retrievedUserA = await getUser(userA.id, scimTokenProviderB);
-			expect(retrievedUserA).toEqual(userA);
-
-			await expect(getUser(userA.id, scimTokenProviderA)).rejects.toThrowError(
-				expect.objectContaining({
-					message: "User not found",
-					body: {
-						detail: "User not found",
-						schemas: ["urn:ietf:params:scim:api:messages:2.0:Error"],
-						status: "404",
-					},
-				}),
-			);
-		});
-
-		it("should return not found for missing users", async () => {
-			const { auth, getSCIMToken } = createTestInstance();
-			const scimToken = await getSCIMToken();
-
-			const getUser = () =>
-				auth.api.getSCIMUser({
-					params: {
-						userId: "missing",
-					},
-					headers: {
-						authorization: `Bearer ${scimToken}`,
-					},
-				});
-
-			await expect(getUser()).rejects.toThrowError(
-				expect.objectContaining({
-					message: "User not found",
-					body: {
-						detail: "User not found",
-						schemas: ["urn:ietf:params:scim:api:messages:2.0:Error"],
-						status: "404",
-					},
-				}),
-			);
-		});
-
-		it("should not allow anonymous access", async () => {
-			const { auth } = createTestInstance();
-
-			const getUser = async () => {
-				await auth.api.getSCIMUser();
-			};
-
-			await expect(getUser()).rejects.toThrow(/SCIM token is required/);
-		});
-	});
-
-	describe("DELETE /scim/v2/Users/:userId", () => {
-		it("should delete an existing user", async () => {
-			const { auth, getSCIMToken } = createTestInstance();
-			const scimToken = await getSCIMToken();
-
-			const newUser = await auth.api.createSCIMUser({
-				body: {
-					userName: "the-username",
-				},
-				headers: {
-					authorization: `Bearer ${scimToken}`,
-				},
-			});
-
-			await auth.api.deleteSCIMUser({
-				params: {
-					userId: newUser.id,
-				},
-				headers: {
-					authorization: `Bearer ${scimToken}`,
-				},
-			});
-
-			const getUser = () =>
-				auth.api.getSCIMUser({
-					params: {
-						userId: newUser.id,
-					},
-					headers: {
-						authorization: `Bearer ${scimToken}`,
-					},
-				});
-
-			await expect(getUser()).rejects.toThrowError(
-				expect.objectContaining({
-					message: "User not found",
-					body: {
-						detail: "User not found",
-						schemas: ["urn:ietf:params:scim:api:messages:2.0:Error"],
-						status: "404",
-					},
-				}),
-			);
-		});
-
-		it("should not allow anonymous access", async () => {
-			const { auth } = createTestInstance();
-
-			const deleteUser = async () => {
-				await auth.api.deleteSCIMUser({
-					params: {
-						userId: "whatever",
-					},
-				});
-			};
-
-			await expect(deleteUser()).rejects.toThrowError(
-				expect.objectContaining({
-					message: "SCIM token is required",
-					body: {
-						detail: "SCIM token is required",
-						schemas: ["urn:ietf:params:scim:api:messages:2.0:Error"],
-						status: "401",
-					},
-				}),
-			);
-		});
-
-		it("should not delete a missing user", async () => {
-			const { auth, getSCIMToken } = createTestInstance();
-			const scimToken = await getSCIMToken();
-
-			const deleteUser = () =>
-				auth.api.deleteSCIMUser({
-					params: {
-						userId: "missing",
-					},
-					headers: {
-						authorization: `Bearer ${scimToken}`,
-					},
-				});
-
-			await expect(deleteUser()).rejects.toThrowError(
-				expect.objectContaining({
-					message: "User not found",
-					body: {
-						detail: "User not found",
-						schemas: ["urn:ietf:params:scim:api:messages:2.0:Error"],
-						status: "404",
-					},
-				}),
-			);
-		});
-	});
-
-	describe("Default SCIM provider", () => {
-		it("should work with a default SCIM provider", async () => {
-			const scimToken = "dGhlLXNjaW0tdG9rZW46dGhlLXNjaW0tcHJvdmlkZXI="; // base64(scimToken:providerId)
-			const { auth } = createTestInstance({
-				defaultSCIM: [
-					{
-						providerId: "the-scim-provider",
-						scimToken: "the-scim-token",
-					},
-				],
-			});
-
-			const createdUser = await auth.api.createSCIMUser({
-				body: {
-					userName: "the-username",
-				},
-				headers: {
-					authorization: `Bearer ${scimToken}`,
-				},
-			});
-
-			expect(createdUser.id).toBeTruthy();
-
-			const user = await auth.api.getSCIMUser({
-				params: {
-					userId: createdUser.id,
-				},
-				headers: {
-					authorization: `Bearer ${scimToken}`,
-				},
-			});
-
-			expect(user).toEqual(createdUser);
-
-			const users = await auth.api.listSCIMUsers({
-				headers: {
-					authorization: `Bearer ${scimToken}`,
-				},
-			});
-
-			expect(users.Resources).toEqual([createdUser]);
-
-			const updatedUser = await auth.api.updateSCIMUser({
-				params: {
-					userId: user.id,
-				},
-				body: {
-					userName: "new-username",
-				},
-				headers: {
-					authorization: `Bearer ${scimToken}`,
-				},
-			});
-
-			expect(updatedUser.userName).toBe("new-username");
-
-			await expect(
-				auth.api.deleteSCIMUser({
-					params: {
-						userId: user.id,
-					},
-					headers: {
-						authorization: `Bearer ${scimToken}`,
-					},
-				}),
-			).resolves.toBe(undefined);
-		});
-
-		it("should reject invalid SCIM tokens", async () => {
-			const { auth } = createTestInstance({
-				defaultSCIM: [
-					{
-						providerId: "the-scim-provider",
-						scimToken: "the-scim-token",
-					},
-				],
-			});
-
-			const createUser = () =>
-				auth.api.createSCIMUser({
-					body: {
-						userName: "the-username",
-					},
-					headers: {
-						authorization: `Bearer invalid-scim-token`,
-					},
-				});
-
-			await expect(createUser()).rejects.toThrow(
-				expect.objectContaining({
-					message: "Invalid SCIM token",
-					body: {
-						detail: "Invalid SCIM token",
-						schemas: ["urn:ietf:params:scim:api:messages:2.0:Error"],
-						status: "401",
-					},
-				}),
-			);
-		});
-	});
 });
