+++ packages/scim/vitest.config.ts
@@ -4,5 +4,6 @@ export default defineProject({
 	test: {
 		clearMocks: true,
 		restoreMocks: true,
+		testTimeout: 10_000,
 	},
 });
